Rename NavigationEventSender->useOnNavigate (2/2)
